gemini-deepgram-elevenlabs-native/pyproject.toml
[project]
name = "gemini-deepgram-elevenlabs-native"
version = "0.1.0"
description = "Voice agent using Gemini LLM, Deepgram STT, and ElevenLabs TTS without frameworks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "plivo>=4.59.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "loguru>=0.7.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "aiohttp>=3.9.0",
    "google-genai>=1.0.0",
    "phonenumbers>=8.13.0",
    "silero-vad>=5.1",
    "torch>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "faster-whisper>=1.2.1",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
